gobuild: rename s3cache.Cache to gobuild.S3Cache
